Add TemplateName type
